requirements.txt
adafruit-circuitpython-dht
streamlit
watchdog
openai
RPi.GPIO
w1thermsensor
